+++ apps/community/urls.py
@@ -25,7 +25,7 @@ from apps.community.views.user.comment_views import (
 )
 from apps.community.views.user.post_create_views import PostCreateAPIView
 from apps.community.views.user.post_delete import PostDeleteAPIView
-from apps.community.views.user.post_like import (
+from apps.community.views.user.post_like_views import (
     PostLikeFalseAPIView,
     PostLikeTrueAPIView,
 )
@@ -76,6 +76,5 @@ urlpatterns = [
     path("posts/<int:post_id>/unlike/", PostLikeFalseAPIView.as_view(), name="post-unlike"),
     path("posts/list", PostListAPIView.as_view(), name="post-list"),
     path("posts/<int:post_id>/update/", UserPostUpdateView.as_view(), name="post-update"),
-    path("api/v1/community/posts/<int:post_id>/", PostDeleteAPIView.as_view(), name="post-delete"),
     path("posts/<int:post_id>/delete/", PostDeleteAPIView.as_view(), name="user-post-delete"),
 ]
